refactor: MarkdownRenderer 통신 타입 정리
